translation_pipeline/pyproject.toml
[tool.poetry]
name = "srbedding"
version = "0.1.0"
description = ""
authors = ["Mitar Perovic <perovicmitar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
datasets = "^2.19.1"
openai = "^1.30.1"
structlog = "^24.1.0"
google-cloud-translate = "^3.15.3"
tqdm = "^4.66.4"
tiktoken = "^0.7.0"
tenacity = "^8.3.0"
cyrtranslit = "^1.1.1"
pandas = "^2.2.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
pytest = "^8.2.0"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
